Fix build warnings
- Fix C1083: Change '../util/tstring.hpp' to '../util/core_types.hpp' in icon_cache_types.hpp
- Fix C5219: Add explicit static_cast<double> in mft_diagnostics.cpp and progress_dialog.hpp
- Fix VCR003: Make ILIsEmpty and global_exception_handler static in UltraFastFileSearch.cpp

Remaining warnings are from third-party code (Windows SDK, WTL, Boost)

+++ UltraFastFileSearch-code/src/cli/mft_diagnostics.cpp
@@ -666,7 +666,7 @@ int benchmark_mft_read(char drive_letter, std::ostream& OS)
     auto end_time = std::chrono::high_resolution_clock::now();
     auto duration = std::chrono::duration_cast<std::chrono::milliseconds>(end_time - start_time);
     double seconds = static_cast<double>(duration.count()) / 1000.0;
-    double mb_per_sec = (seconds > 0) ? (bytes_read_total / (1024.0 * 1024.0)) / seconds : 0;
+    double mb_per_sec = (seconds > 0) ? (static_cast<double>(bytes_read_total) / (1024.0 * 1024.0)) / seconds : 0;
 
     CloseHandle(volume_handle);
